:card_file_box: add: 게시글 컴포넌트 추가
